Move diff suppress funcs used by a single services into that service's package (#9962)
[upstream:167ad5f2228223e948b669d64767dbff650f1fc3]
