+++ google-beta/services/compute/resource_compute_region_disk.go
@@ -166,7 +166,7 @@ Applicable only for bootable disks.`,
 				Optional:         true,
 				Deprecated:       "`interface` is deprecated and will be removed in a future major release. This field is no longer used and can be safely removed from your configurations; disk interfaces are automatically determined on attachment.",
 				ForceNew:         true,
-				DiffSuppressFunc: tpgresource.AlwaysDiffSuppress,
+				DiffSuppressFunc: AlwaysDiffSuppress,
 				Description:      `Specifies the disk interface to use for attaching this disk, which is either SCSI or NVME. The default is SCSI.`,
 				Default:          "SCSI",
 			},
