+++ google-beta/services/pubsub/resource_pubsub_subscription_test.go
@@ -367,6 +367,39 @@ func TestAccPubsubSubscription_pollOnCreate(t *testing.T) {
 	})
 }
 
+func TestUnitPubsubSubscription_IgnoreMissingKeyInMap(t *testing.T) {
+	cases := map[string]struct {
+		Old, New           string
+		Key                string
+		ExpectDiffSuppress bool
+	}{
+		"missing key in map": {
+			Old:                "",
+			New:                "v1",
+			Key:                "x-goog-version",
+			ExpectDiffSuppress: true,
+		},
+		"different values": {
+			Old:                "v1",
+			New:                "v2",
+			Key:                "x-goog-version",
+			ExpectDiffSuppress: false,
+		},
+		"same values": {
+			Old:                "v1",
+			New:                "v1",
+			Key:                "x-goog-version",
+			ExpectDiffSuppress: false,
+		},
+	}
+
+	for tn, tc := range cases {
+		if pubsub.IgnoreMissingKeyInMap(tc.Key)("push_config.0.attributes."+tc.Key, tc.Old, tc.New, nil) != tc.ExpectDiffSuppress {
+			t.Fatalf("bad: %s, '%s' => '%s' expect %t", tn, tc.Old, tc.New, tc.ExpectDiffSuppress)
+		}
+	}
+}
+
 func testAccPubsubSubscription_emptyTTL(topic, subscription string) string {
 	return fmt.Sprintf(`
 resource "google_pubsub_topic" "foo" {
