+++ google-beta/tpgresource/utils_test.go
@@ -227,93 +227,6 @@ func TestConvertStringMap(t *testing.T) {
 	}
 }
 
-func TestIpCidrRangeDiffSuppress(t *testing.T) {
-	cases := map[string]struct {
-		Old, New           string
-		ExpectDiffSuppress bool
-	}{
-		"single ip address": {
-			Old:                "10.2.3.4",
-			New:                "10.2.3.5",
-			ExpectDiffSuppress: false,
-		},
-		"cidr format string": {
-			Old:                "10.1.2.0/24",
-			New:                "10.1.3.0/24",
-			ExpectDiffSuppress: false,
-		},
-		"netmask same mask": {
-			Old:                "10.1.2.0/24",
-			New:                "/24",
-			ExpectDiffSuppress: true,
-		},
-		"netmask different mask": {
-			Old:                "10.1.2.0/24",
-			New:                "/32",
-			ExpectDiffSuppress: false,
-		},
-		"add netmask": {
-			Old:                "",
-			New:                "/24",
-			ExpectDiffSuppress: false,
-		},
-		"remove netmask": {
-			Old:                "/24",
-			New:                "",
-			ExpectDiffSuppress: false,
-		},
-	}
-
-	for tn, tc := range cases {
-		if tpgresource.IpCidrRangeDiffSuppress("ip_cidr_range", tc.Old, tc.New, nil) != tc.ExpectDiffSuppress {
-			t.Fatalf("bad: %s, '%s' => '%s' expect %t", tn, tc.Old, tc.New, tc.ExpectDiffSuppress)
-		}
-	}
-}
-
-func TestRfc3339TimeDiffSuppress(t *testing.T) {
-	cases := map[string]struct {
-		Old, New           string
-		ExpectDiffSuppress bool
-	}{
-		"same time, format changed to have leading zero": {
-			Old:                "2:00",
-			New:                "02:00",
-			ExpectDiffSuppress: true,
-		},
-		"same time, format changed not to have leading zero": {
-			Old:                "02:00",
-			New:                "2:00",
-			ExpectDiffSuppress: true,
-		},
-		"different time, both without leading zero": {
-			Old:                "2:00",
-			New:                "3:00",
-			ExpectDiffSuppress: false,
-		},
-		"different time, old with leading zero, new without": {
-			Old:                "02:00",
-			New:                "3:00",
-			ExpectDiffSuppress: false,
-		},
-		"different time, new with leading zero, oldwithout": {
-			Old:                "2:00",
-			New:                "03:00",
-			ExpectDiffSuppress: false,
-		},
-		"different time, both with leading zero": {
-			Old:                "02:00",
-			New:                "03:00",
-			ExpectDiffSuppress: false,
-		},
-	}
-	for tn, tc := range cases {
-		if tpgresource.Rfc3339TimeDiffSuppress("time", tc.Old, tc.New, nil) != tc.ExpectDiffSuppress {
-			t.Errorf("bad: %s, '%s' => '%s' expect DiffSuppress to return %t", tn, tc.Old, tc.New, tc.ExpectDiffSuppress)
-		}
-	}
-}
-
 func TestGetProject(t *testing.T) {
 	cases := map[string]struct {
 		ResourceConfig  map[string]interface{}
